Modifying date validation

+++ src/main/java/model/api/ListItem.java
@@ -1,6 +1,5 @@
 package model.api;
 
-import java.util.List;
 import com.google.gson.annotations.SerializedName;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -9,14 +8,14 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class ListItem{
+public class ListItem {
 
-	@SerializedName("dt")
-	private int dt;
+    @SerializedName("dt")
+    private long dt;
 
-	@SerializedName("main")
-	private Main main;
+    @SerializedName("main")
+    private Main main;
 
-	@SerializedName("wind")
-	private Wind wind;
+    @SerializedName("wind")
+    private Wind wind;
 }
